eat(biome_js_analyze): extracts class member references in a service)
