+++ crates/biome_js_analyze/src/lint/style/use_readonly_class_properties.rs
@@ -1,5 +1,5 @@
 use crate::JsRuleAction;
-use crate::services::class_member_references::{AnyPropertyMember, ClassMemberReference, ClassMemberReferences, SemanticClass};
+use crate::services::semantic_class::{AnyPropertyMember, ClassMemberReference, ClassMemberReferences, SemanticClass};
 use biome_analyze::{FixKind, Rule, RuleDiagnostic, RuleSource, context::RuleContext, declare_lint_rule};
 use biome_console::markup;
 use biome_js_factory::make;
