+++ crates/biome_js_analyze/src/services/mod.rs
@@ -4,4 +4,4 @@ pub mod manifest;
 pub mod module_graph;
 pub mod semantic;
 pub mod typed;
-pub mod class_member_references;
+pub mod semantic_class;
