+++ crates/biome_js_analyze/src/services/semantic_class.rs
